Update the documentation for adding custom icons to Icons block
Fixed the documentation about adding custom icons to the Icons block. The old did not give specific instructions about the naming convention needed for adding custom SVG icons. This documentation mentions that the SVGs need to have the same name as the json key that references that SVG. If you followed the instructions from the old doc, your icons would not load in the Icons block.